pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "credx-ai"
version = "1.0.0"
description = "Your Personal Career Trajectory Engine - AI-powered job matching platform"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    {name = "Nitin RC", email = "nitinrc264@example.com"},
]
keywords = ["ai", "job-matching", "career", "resume-parser", "machine-learning"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "Flask>=2.3.0",
    "Flask-Cors>=4.0.0",
    "gunicorn",
    "pandas>=2.0.0",
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "torchaudio>=2.0.0",
    "PyMuPDF>=1.23.0",
    "google-generativeai>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/nitinc264/CredX-AI"
Repository = "https://github.com/nitinc264/CredX-AI"
Issues = "https://github.com/nitinc264/CredX-AI/issues"
Demo = "https://credx-ai.onrender.com/"

[tool.setuptools]
py-modules = []

[tool.setuptools.packages.find]
where = ["src"]
include = ["credx_ai", "credx_ai.*"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
